Remove queries module
