+++ packages/miniplex-core/src/Query.ts
@@ -1,5 +1,17 @@
 import { Bucket } from "@miniplex/bucket"
 import { With, Without } from "./types"
+import { ArchetypeQuery } from "./types"
+
+export const normalizeComponents = (components: any[]) => [
+  ...new Set(components.sort().filter((c) => !!c && c !== ""))
+]
+
+export const normalizeQuery = (query: ArchetypeQuery<any, any>) =>
+  ({
+    with: query.with !== undefined ? normalizeComponents(query.with) : [],
+    without:
+      query.without !== undefined ? normalizeComponents(query.without) : []
+  } as typeof query)
 
 export type QueryConfiguration<E> = {
   with: (keyof E)[]
