+++ packages/miniplex-core/src/index.ts
@@ -1,4 +1,3 @@
 export * from "@miniplex/bucket"
-export * from "./queries"
 export * from "./types"
 export * from "./World"
